backend/requirements.txt
flask
pymongo
flask_jwt_extended
python-dotenv
flask_smorest
flask_cors
pydantic[email]
hashlib
bcrypt